Extract Dokploy target CLI commands
